Add files via upload
Introduction and Problem Statement
The Ames Housing dataset, which considers the features and price points of different properties in Ames, Ohio was the basis for this analysis.

In this analysis, I carried out some steps to bring meaning to the different features in the dataset. I began by formatting the data, which involved uploading it on Google Colab. The next steps involved data cleaning, Exploratory Data Analysis, data visualization and splitting trhe dataset into testing and training.

The aim of this analysis is to showcase the relationship that different elements in the dataset has with SalePrice, which is the target variable in this analysis.